backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
sentence-transformers
faiss-cpu
pydantic
